requirements.txt
# Core
pandas>=2.1
numpy>=1.25
psycopg2-binary>=2.9
SQLAlchemy>=2.0

# dbt (core + postgres adapter)
dbt-core==1.5.0
dbt-postgres==1.5.0

# orchestration
prefect==2.6.0

# Streamlit
streamlit>=1.25

# Recommender
mlxtend>=0.21.0

# Helpers & utils
python-dotenv
sqlglot
requests
pyarrow
kaggle

